Add global types and utils
